Change compressed_size type for *really* big fonts

+++ src/epd_driver/include/epd_internals.h
@@ -118,7 +118,7 @@ typedef struct {
   uint16_t advance_x;        ///< Distance to advance cursor (x axis)
   int16_t left;             ///< X dist from cursor pos to UL corner
   int16_t top;              ///< Y dist from cursor pos to UL corner
-  uint16_t compressed_size; ///< Size of the zlib-compressed font data.
+  uint32_t compressed_size; ///< Size of the zlib-compressed font data.
   uint32_t data_offset;     ///< Pointer into EpdFont->bitmap
 } EpdGlyph;
